requirements.txt
astroid==3.3.8
attrs==25.1.0
dill==0.3.9
gitdb==4.0.12
GitPython==3.1.44
iniconfig==2.0.0
isort==6.0.0
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pylint==3.3.4
pytest==8.3.4
pytest-subtests==0.14.1
smmap==5.0.2
tomlkit==0.13.2
